+++ pingoverride/ping.py
@@ -24,9 +24,13 @@ SOFTWARE.
 
 import discord
 import logging
+import asyncio
+import random
 
 from redbot.core import Config, commands
 from redbot.core.utils.chat_formatting import bold, box, pagify
+from redbot.core.utils.predicates import ReactionPredicate
+from redbot.core.utils.menus import start_adding_reactions
 
 log = logging.getLogger("red.kreusada.pingoverride")
 
@@ -36,7 +40,9 @@ class PingOverride(commands.Cog):
     Custom ping message.
     """
 
-    __author__ = ["Kreusada", ]
+    __author__ = [
+        "Kreusada",
+    ]
     __version__ = "1.7.0"
 
     def __init__(self, bot):
@@ -44,9 +50,7 @@ class PingOverride(commands.Cog):
         self.config = Config.get_conf(
             self, identifier=59365034743, force_registration=True
         )
-        self.config.register_global(
-            response="Pong.", reply=False, mention=True, embed=False
-        )
+        self.config.register_global(response=[], reply=False, mention=True, embed=False)
 
     def format_help_for_context(self, ctx: commands.Context) -> str:
         """Thanks Sinbad."""
@@ -76,6 +80,28 @@ class PingOverride(commands.Cog):
             }
         return match.format(**mapping)
 
+    async def enum(self, ctx, message_list):
+        msg_list = [await self.converter(ctx, x, False) for x in message_list]
+        pre_processed = box(
+            "\n".join(f"+ {c+1}: {self.shorten(v)}" for c, v in enumerate(msg_list)),
+            lang="diff",
+        )
+        message = f"The following responses have been set! {pre_processed}"
+
+        response = await self.config.response()
+        for z in message_list:
+            response.append(z)
+        await self.config.response.set(response)
+
+        return await ctx.send(message)
+
+    @staticmethod
+    def shorten(text):
+        if len(text) > 30:
+            return text[:30] + "..."
+        else:
+            return text
+
     @commands.is_owner()
     @commands.group()
     async def pingset(self, ctx: commands.Context):
@@ -101,27 +127,61 @@ class PingOverride(commands.Cog):
     @pingset.command()
     async def settings(self, ctx: commands.Context):
         """Get the settings for the ping command."""
-        response = await self.config.response()
         cross = "\N{CROSS MARK}"
         check = "\N{WHITE HEAVY CHECK MARK}"
+
+        response = await self.config.response()
+        reply = await self.config.reply()
+        mention = await self.config.mention()
+        embed = await self.config.embed()
+
+        if not response:
+            pre_processed = "+ Pong."
+        else:
+            message_list = [await self.converter(ctx, x, False) for x in response]
+            pre_processed = "\n".join(
+                f"+ {c+1}: {self.shorten(v)}" for c, v in enumerate(message_list)
+            )
+
+        pre_processed = box(pre_processed, lang="diff")
+
         if await ctx.embed_requested():
             embed = discord.Embed(
-                title=f"Settings for {ctx.bot.user.name}",
-                color=await ctx.embed_colour()
+                title=f"Ping Settings for {ctx.bot.user.name}",
+                color=await ctx.embed_colour(),
             )
-            embed.add_field(name="response", value=await self.converter(ctx, response, False), inline=False)
-            embed.add_field(name="Replies", value=check if await self.config.reply() else cross, inline=False)
-            embed.add_field(name="Reply mentions", value=check if await self.config.mention() else cross, inline=False)
-            embed.add_field(name="Embeds", value=check if await self.config.embed() else cross, inline=False)
+
+            embed.add_field(
+                name="Replies", value=check if reply else cross, inline=True
+            )
+
+            if reply:
+                embed.add_field(
+                    name="Reply mentions",
+                    value=check if mention else cross,
+                    inline=True,
+                )
+
+            embed.add_field(name="Embeds", value=check if embed else cross, inline=True)
+            embed.add_field(name="Responses", value=pre_processed, inline=False)
+            embed.set_footer(text=f"See {ctx.clean_prefix}pingset regex, for information on response regex.")
+
             await ctx.send(embed=embed)
         else:
-            text = (
-                f"{bold('Response:')} {await self.converter(ctx, response, False)}\n"
-                f"{bold('Replies:')} {await self.config.reply()}\n"
-                f"{bold('Reply mentions:')} {await self.config.mention()}\n"
-                f"{bold('Use embeds:')} {await self.config.embed()}"
+            await ctx.send("I need to be able to send embeds.")
+
+    @pingset.command()
+    async def regex(self, ctx: commands.Context):
+        """Get information on the types of ping regex."""
+        description = "`{latency}`: Bot Latency\n`{display}`: Author's Display Name"
+        if await ctx.embed_requested():
+            await ctx.send(
+                embed=discord.Embed(
+                    description=description, color=await ctx.embed_colour()
+                )
             )
-            await ctx.send(text)
+        else:
+            await ctx.send(description)
 
     @pingset.command()
     async def embed(self, ctx: commands.Context, true_or_false: bool):
@@ -137,7 +197,7 @@ class PingOverride(commands.Cog):
 
     @pingset.command()
     @commands.guild_only()
-    async def message(self, ctx: commands.Context, *, response: str):
+    async def message(self, ctx: commands.Context, *, message: str):
         """
         Set your custom ping message.
 
@@ -148,24 +208,70 @@ class PingOverride(commands.Cog):
         Example Usage:
         `[p]pingset message Hello {display}! My latency is {latency} ms.`
         """
-        await self.config.response.set(response)
-        msg = await self.converter(ctx, response, False)
-        embed = await self.config.embed()
-        if embed:
-            embed = discord.Embed(description=msg, color=await ctx.embed_colour())
-            await ctx.send(
-                content=f"Running `{ctx.clean_prefix}ping` will now respond with...",
-                embed=embed,
+
+        msg = await ctx.send(
+            "Would you like to add any other responses, to be chosen at random?"
+        )
+        pred = ReactionPredicate.yes_or_no(msg, ctx.author)
+        start_adding_reactions(msg, ReactionPredicate.YES_OR_NO_EMOJIS)
+
+        try:
+            await self.bot.wait_for("reaction_add", check=pred, timeout=30)
+        except asyncio.TimeoutError:
+            await self.config.response.clear()
+            response = await self.config.response()
+            response.append(message)
+            await self.config.response.set(response)
+            return await ctx.send(
+                "You took too long to answer, I'll stick to this one response!"
             )
+
+        if pred.result:
+            await ctx.send("Okay, let's add some random responses.")
+            await self.config.response.clear()
+
+            message_list = [message]
+            response = await self.config.response()
+
+            while True:
+
+                if len(message_list) > 9:
+                    await ctx.send("You've reached the maximum number of responses!")
+                    return await self.enum(ctx, message_list)
+
+                await ctx.send("Add a random response:")
+
+                def check(x):
+                    return x.author == ctx.author and x.channel == ctx.channel
+
+                try:
+                    add_response = await self.bot.wait_for(
+                        "message", timeout=50, check=check
+                    )
+                except asyncio.TimeoutError:
+                    return await ctx.send("Timed out. No changes have been made.")
+
+                if add_response.content.lower().startswith(("exit()", "stop()")):
+                    await ctx.send("Ended!")
+                    return await self.enum(ctx, message_list)
+                else:
+                    message_list.append(add_response.content)
+
         else:
-            await ctx.send(
-                f"Running `{ctx.clean_prefix}ping` will now respond with... {box(msg, lang='yaml')}"
-            )
+            await self.config.response.clear()
+            response = await self.config.response()
+            response.append(message)
+            await self.config.response.set(response)
+            return await ctx.send("Ok, I'll stick to this one response!")
 
     @commands.command()
     async def ping(self, ctx: commands.Context):
         """Pong. Or not?"""
         resp = await self.config.response()
+        if not resp:
+            resp = "Pong."
+        else:
+            resp = random.choice(resp)
         reply = await self.config.reply()
         mention = await self.config.mention()
         embed = await self.config.embed()